+++ nativeApp/src/redux/logs.ts
@@ -2,7 +2,7 @@ import { Reducer } from 'redux';
 import { createAction } from 'redux-actions';
 import { assoc } from 'ramda';
 import { VolunteerLog } from '../../../api/src/models/types';
-import { VolunteerLogs } from '../lib/api';
+import { VolunteerLogs } from '../api';
 import { State, LogsState } from './types';
